requirements.txt
appdirs==1.4.3
Django==1.10.6
django-appconf==1.0.2
django-compressor==2.1.1
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
